+++ lib/mojito.js
@@ -40,33 +40,8 @@ var debug = require('debug')('mojito'),
     liblogger = require('./logger'),
     libmiddleware = require('./middleware'),
     libstore = require('./store'),
+    extend = require('./util').extend,
     Mojito;
-
-
-//  ----------------------------------------------------------------------------
-//  Util
-//  ----------------------------------------------------------------------------
-// If more helpers are needed, they could potentially be moved to their own
-// `util.js` module later.
-
-/**
-Copied from modown-yui.utils
-**/
-function extend(obj) {
-    Array.prototype.slice.call(arguments, 1).forEach(function (source) {
-        var key;
-
-        if (!source) { return; }
-
-        for (key in source) {
-            if (source.hasOwnProperty(key)) {
-                obj[key] = source[key];
-            }
-        }
-    });
-
-    return obj;
-}
 
 
 //  ----------------------------------------------------------------------------
