+++ lib/util.js
@@ -11,6 +11,37 @@
 var libpath = require('path');
 
 /**
+Extends object with properties from other objects.
+
+    var a = { foo: 'bar' }
+      , b = { bar: 'baz' }
+      , c = { baz: 'xyz' };
+
+    utils.extends(a, b, c);
+    // a => { foo: 'bar', bar: 'baz', baz: 'xyz' }
+
+@method extend
+@param {Object} obj the receiver object to be extended
+@param {Object*} supplier objects
+@return {Object} The extended object
+**/
+function extend(obj) {
+    Array.prototype.slice.call(arguments, 1).forEach(function (source) {
+        var key;
+
+        if (!source) { return; }
+
+        for (key in source) {
+            if (source.hasOwnProperty(key)) {
+                obj[key] = source[key];
+            }
+        }
+    });
+
+    return obj;
+}
+
+/**
 Produces a normalized web path by joining all the parts and normalizing the
 filesystem-like path into web compatible url. This is useful when you have to
 generate urls based on filesystem path where unix uses `/` and windows uses `\\`.
@@ -31,8 +62,13 @@ paths.
 @param {Array|String*} url the list of parts to be joined and normalized
 @return {String} The joined and normalized url
 **/
-exports.webpath = function (url) {
+function webpath(url) {
     var args = [].concat.apply([], arguments),
         parts = libpath.join.apply(libpath, args).split(libpath.sep);
     return parts.join('/');
+}
+
+exports = module.exports = {
+    extend: extend,
+    webpath: webpath
 };
